pipeline/application/batch/orchestration/pyproject.toml
[project]
name = "orchestration"
version = "0.1.0"
description = "Airflow DAGs for batch transaction processing orchestration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Note: These are installed via Dockerfile (pip), not uv
    # Kept here for documentation and local development type checking
    "apache-airflow==3.1.5",
    "apache-airflow-providers-docker==3.14.0",
]

[dependency-groups]
dev = [
    "pytest >= 8",
    "pytest-cov >= 5",
    "coverage[toml] >= 7.6.10",
    "poethepoet >= 0.32.1"
]

lint = [
    "pre-commit>=4.2.0",
    "pyright>=1.1.391",
    "ruff>=0.9.1",
    "pydocstyle>=6.3.0",
]

[tool.uv]
cache-dir = "./.cache"
default-groups = ["dev", "lint"]

[tool.coverage.paths]
source = ["."]

[tool.coverage.run]
branch = true
data_file = ".cache/coverage"
source = ["."]
omit = ["tests/*", ".venv/*", "*/__main__.py", "dags/*"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers --pdbcls=tests:Debugger -r sxX --cov=. --cov-report=html --cov-report=term-missing:skip-covered"
cache_dir = ".cache"

[tool.ruff]
cache-dir = ".cache/ruff"
line-length = 120
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle
    "F",    # Pyflakes
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "I",    # isort
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.pyright]
include = ["dags"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    ".venv",
    ".cache",
]
ignore = ["tests"]
defineConstant = { DEBUG = true }
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.13"
pythonPlatform = "Linux"
venvPath = "../../../.."
venv = ".venv"

[tool.poe.tasks.types]
help = "Run the type checker"
cmd  = "uv run pyright"

[tool.poe.tasks.test]
help = "Run unit tests"
shell = """
if test -e .cache/v/cache/lastfailed; then uv run pytest tests --last-failed --exitfirst; fi &
rm -rf .cache/v/cache/lastfailed &
uv run pytest
"""

[tool.poe.tasks.lint]
help = "Run all checks on the code base"
shell = """
uv run pydocstyle
uv run ruff check --fix .
uv run ruff format --check .
uv run pyright
"""

[tool.poe.tasks.linter-fix]
help ="Run Linter in fix mode."
shell = """
uv run ruff format .
uv run ruff check --select I --fix .
"""

[tool.poe.tasks.style]
help = "Validate code style"
shell = """
uv run ruff check --select I --fix
uv run ruff format .
"""

[tool.poe.tasks.check]
help     = "Run all checks on the code base"
sequence = ["style", "types", "lint"]

[tool.poe.tasks.pre-commit]
help = "Run pre-commit hooks on all files"
cmd  = "uv run pre-commit run --all-files"

[tool.poe.tasks.clean]
help = "Remove all generated and temporary files"
shell  = """
uv cache clean
rm -rf *.spec dist build .eggs *.egg-info .install .cache .coverage htmlcov .mypy_cache .pytest_cache .ruff_cache &
find . -type d -name '__pycache__' -exec rm -rf {} +
"""
